+++ src/main/java/com/yuuko/core/Cache.java
@@ -9,7 +9,6 @@ import net.dv8tion.jda.core.entities.User;
 import org.discordbots.api.client.DiscordBotListAPI;
 
 import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
 
 public class Cache {
@@ -19,7 +18,6 @@ public class Cache {
     public static List<Module> MODULES;
     public static List<String> SETTINGS;
     public static DiscordBotListAPI BOT_LIST;
-    public static LinkedList<String> LAST_THIRTEEN;
     public static String LATEST_INFO;
     public static String[] STANDARD_STRINGS;
     public static final HashMap<Long, List<SearchResult>> audioSearchResults = new HashMap<>();