+++ src/main/java/com/yuuko/core/Yuuko.java
@@ -29,7 +29,6 @@ import javax.security.auth.login.LoginException;
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
 
@@ -122,10 +121,6 @@ public class Yuuko {
             Cache.SETTINGS = settingsList;
 
             Cache.LATEST_INFO = "";
-            Cache.LAST_THIRTEEN = new LinkedList<>();
-            for(int i = 0; i < 13; i++) {
-                Cache.LAST_THIRTEEN.add("");
-            }
 
             ScheduleHandler.registerJob(new FiveSecondlyJob());
             ScheduleHandler.registerJob(new ThirtySecondlyJob());